cleaned up senddm
